Move directory function from store → store/dirnames
This is to break the cycle that store depends on limayaml, and networks
depends on store, so limayaml cannot depend on networks to allow network
name validation.
