+++ pkg/store/store.go
@@ -7,50 +7,12 @@ import (
 
 	"github.com/containerd/containerd/identifiers"
 	"github.com/lima-vm/lima/pkg/limayaml"
-	"github.com/lima-vm/lima/pkg/store/filenames"
+	"github.com/lima-vm/lima/pkg/store/dirnames"
 )
-
-// DotLima is a directory that appears under the home directory.
-const DotLima = ".lima"
-
-// LimaDir returns the abstract path of `~/.lima` (or $LIMA_HOME, if set).
-//
-// NOTE: We do not use `~/Library/Application Support/Lima` on macOS.
-// We use `~/.lima` so that we can have enough space for the length of the socket path,
-// which can be only 104 characters on macOS.
-func LimaDir() (string, error) {
-	dir := os.Getenv("LIMA_HOME")
-	if dir == "" {
-		homeDir, err := os.UserHomeDir()
-		if err != nil {
-			return "", err
-		}
-		dir = filepath.Join(homeDir, DotLima)
-	}
-	return dir, nil
-}
-
-// LimaConfigDir returns the path of the config directory, $LIMA_HOME/_config.
-func LimaConfigDir() (string, error) {
-	limaDir, err := LimaDir()
-	if err != nil {
-		return "", err
-	}
-	return filepath.Join(limaDir, filenames.ConfigDir), nil
-}
-
-// LimaNetworksDir returns the path of the networks log directory, $LIMA_HOME/_networks.
-func LimaNetworksDir() (string, error) {
-	limaDir, err := LimaDir()
-	if err != nil {
-		return "", err
-	}
-	return filepath.Join(limaDir, filenames.NetworksDir), nil
-}
 
 // Instances returns the names of the instances under LimaDir.
 func Instances() ([]string, error) {
-	limaDir, err := LimaDir()
+	limaDir, err := dirnames.LimaDir()
 	if err != nil {
 		return nil, err
 	}
@@ -74,7 +36,7 @@ func InstanceDir(name string) (string, error) {
 	if err := identifiers.Validate(name); err != nil {
 		return "", err
 	}
-	limaDir, err := LimaDir()
+	limaDir, err := dirnames.LimaDir()
 	if err != nil {
 		return "", err
 	}
